pyproject.toml
[project]
name = "PyHub"
version = "0.1.3"
description = "A fun hub for different kinds of utilities"
requires-python = ">=3.12"

authors = [
    { name = "Lakshya" }
]

readme = "README.md"
license = { text = "CC BY-NC 4.0" }

keywords = ["utilities", "tools", "hub", "fun", "python"]

classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent"
]

dependencies = [
    "plotext",
    "numpy",
    "pywebview"
]

[project.scripts]
pyhub = "PyHub.main:main"
